tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/navigator_user_agent.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(t => {
    assert_false("getUserAgent" in navigator);
  }, "navigator.getUserAgent() is not available in non-secure contexts.");
</script>
