Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/navigator_user_agent.tentative-expected.txt


PASS navigator.getUserAgent() is not available in non-secure contexts.
